requirements.txt
flask
flask-socketio
azure-cognitiveservices-speech
requests
python-dotenv
firebase-admin